requirements.txt
openai-whisper
openai>=1.0.0
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
python-Levenshtein>=0.21.0
scipy>=1.10.0
python-dotenv>=1.0.0

# AssemblyAI dependencies
requests>=2.31.0